requirements.txt
streamlit
firebase-admin
streamlit-drawable-canvas
Pillow
plotly
pandas